Move headers to a subdir
For clearer scoping when using the library in an application
