+++ usbplusplus/byteorder.hpp
