+++ usbplusplus/usblangids.hpp
