+++ usbplusplus/ustring.hpp
